backend/requirements.txt
flask>=2.3
flask-cors>=4.0
google-genai>=0.18.0
language-tool-python>=2.9
python-dotenv>=1.0
requests>=2.31
